test: test 가독성위해 테스트 분리 및 중복제거 처리
